requirements.txt
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Markdown==3.1.1
simplejson==3.16.0
PyJWT==1.7.1
uritemplate==3.0.0
django-ipware==2.1.0
